feat: rewrite all input form parts components to funtional
