+++ my-react-app/src/components/inputComponents/inputSelect/inputSelect.tsx
@@ -1,0 +1,33 @@
+import { InputSelectInterface } from 'models/models';
+import React from 'react';
+
+export function InputSelect({ valid, inputRef }: InputSelectInterface) {
+  return (
+    <div className="form__country form-block">
+      <label className="form__country--label form-label" htmlFor="form__country">
+        Country:
+      </label>
+      <select
+        className="form__country--select form-select"
+        id="form__country"
+        ref={inputRef}
+        defaultValue=""
+      >
+        <option className="form__country--option" disabled></option>
+        <option className="form__country--option" value="USA">
+          USA
+        </option>
+        <option className="form__country--option" value="Canada">
+          Canada
+        </option>
+        <option className="form__country--option" value="Mexico">
+          Mexico
+        </option>
+        <option className="form__country--option" value="Germany">
+          Germany
+        </option>
+      </select>
+      {!valid && <span className="form__country--span error-span">Error! No country selected</span>}
+    </div>
+  );
+}
