+++ my-react-app/src/components/inputComponents/searchInput/SearchInput.css
